+++ R/OMOPGenerics-package.R
@@ -6,5 +6,6 @@
 #' @importFrom rlang :=
 #' @importFrom rlang .data
 #' @importFrom rlang .env
+#' @importFrom rlang %||%
 ## usethis namespace: end
 NULL