+++ tests/testthat/test-logger.R
@@ -43,11 +43,9 @@ test_that("test sql logging", {
 
   sqlPath <- file.path(tempdir(), "sql_files")
   dir.create(sqlPath)
-  explainPath <- file.path(tempdir(), "explain_files")
-  dir.create(explainPath)
 
   options(omopgenerics.log_sql_path = sqlPath)
-  options(omopgenerics.log_sql_explain_path = explainPath)
+  options(omopgenerics.log_sql_explain = TRUE)
 
   cdm$drug_exposure <- cdm$drug_exposure |>
     dplyr::left_join(
@@ -58,10 +56,8 @@ test_that("test sql logging", {
     dplyr::compute(name = "drug_exposure")
 
   res1 <- summariseLogSqlPath()
-  res2 <- summariseLogExplainPath()
 
   dropSourceTable(cdm = cdm, name = dplyr::everything())
   cdmDisconnect(cdm = cdm)
   unlink(x = sqlPath, recursive = TRUE)
-  unlink(x = explainPath, recursive = TRUE)
 })